- fixed ETCX support for RGB-only textures

+++ scripts/etcx/etcx.py
@@ -15,6 +15,9 @@ def _read_file(filename):
 class Etcx:
 	
 	Version = 1
+	HEADER_HAS_ALPHA_BIT 0x1
+	HEADER_IS_ZLIB_COMPRESSED_BIT 0x2
+	
 	HEADER_SIZE = 13 * 4
 	# etcpak has the w/h order reversed for some reason
 	WIDTH_HEADER_OFFSET = 7 * 4
@@ -25,26 +28,33 @@ class Etcx:
 		if not os.path.exists(etc1):
 			return "ERROR! File '%s' does not exist!" % etc1
 		if not os.path.exists(etc1a):
-			return "ERROR! File '%s' does not exist!" % etc1a
+			etc1a = None
 		if zlibCompressionLevel < 0:
 			zlibCompressionLevel = 0
 		elif zlibCompressionLevel > 9:
 			zlibCompressionLevel = 9
+		flags = 0x0
 		header, etc1_data = _read_file(etc1)
-		dummy_header, etc1a_data = _read_file(etc1a)
+		if etc1a != None:
+			flags = Etcx.HEADER_HAS_ALPHA_BIT
+			dummy_header, etc1a_data = _read_file(etc1a)
 		f = open(etcx, "wb")
 		# header
 		f.write("ETCX")
-		flags = 0x1
 		if zlibCompressionLevel > 0:
-			flags = flags | 0x2
+			flags |= Etcx.HEADER_IS_ZLIB_COMPRESSED_BIT
 		f.write(struct.pack(UINT_PACK_FORMAT, flags))
 		f.write(header[Etcx.WIDTH_HEADER_OFFSET : Etcx.WIDTH_HEADER_OFFSET + 4])
 		f.write(header[Etcx.HEIGHT_HEADER_OFFSET : Etcx.HEIGHT_HEADER_OFFSET + 4])
-		size = len(etc1_data) + len(etc1a_data)
+		size = len(etc1_data)
+		if etc1a != None:
+			size += len(etc1a_data)
 		f.write(struct.pack(UINT_PACK_FORMAT, size))
 		if zlibCompressionLevel > 0:
-			data = zlib.compress(etc1_data + etc1a_data, zlibCompressionLevel)
+			if etc1a != None:
+				data = zlib.compress(etc1_data + etc1a_data, zlibCompressionLevel)
+			else:
+				data = zlib.compress(etc1_data, zlibCompressionLevel)
 			f.write(struct.pack(UINT_PACK_FORMAT, len(data)))
 			f.write(data)
 		else:
@@ -52,9 +62,11 @@ class Etcx:
 			# ETC1
 			f.write(etc1_data)
 			# ETC1 alpha
-			f.write(etc1a_data)
+			if etc1a != None:
+				f.write(etc1a_data)
 		# done
 		f.close()
 		os.remove(etc1)
-		os.remove(etc1a)
+		if etc1a != None:
+			os.remove(etc1a)
 		return "File '%s' has been successfully merged." % etcx
